Add recipe for Chicken ChickenCurry
Add a new recipe for Indian style Chicken Curry to the recipes list.
The page structure is the same as before with a picture followed by
a description and the cooking procedure.

The new recipe gets linked to from the recipes index page.
As we have more than one link on the index page now, we have to structure
the links list as an unordered list.